==> R/00_data_cleaning.qmd <==
---
title: "Data Programming - Group Assignment"
format: html
editor: source
---

# DATA CLEANING PROCESS

In the following chunk, I will load the packages that I will use for the data cleaning process.

```{r}
#| messsages: false
#| warning: false

library(gert)

library(tidyverse)
library(stringi) 
library(rvest)
```

Before proceeding, I will first load all the data files:

```{r}
#| messsages: false
#| warning: false

election_data <- read_csv("data/raw/datos_elecciones_brutos.csv")
cod_mun <- read_csv("data/raw/cod_mun.csv")
surveys <- read_csv("data/raw/historical_surveys.csv")
abbrev <- read_csv("data/raw/siglas.csv")
```

Now I will take a look to each of them, clean them one by one and join them if necessary so we can work with as few data files as possible.

## election_data

First I will use `pivot_longer( )` to create a new variable that stores the name of each party, and another one that includes the number of votes for each of them:

```{r}
election_data <- election_data |> 
  pivot_longer(cols = -(tipo_eleccion:votos_candidaturas), 
               names_to = "partido",
               values_to = "num_votos")           
```

Then, I remove the `tipo_eleccion`variable, because all of them are parliamentary elections, so it does not provide any useful information. I do the same with the `vuelta` variable, because parliamentary election in Spain only have one round (in contrast to ballotage systems).

```{r}
election_data <- election_data |> 
  select(-tipo_eleccion) |> 
  select(-vuelta)
```

Now I will select only those parties we are interested in:

-   PARTIDO SOCIALISTA OBRERO ESPAÑOL (beware: it has/had federations - branches - with some other name).
-   PARTIDO POPULAR
-   CIUDADANOS (caution: has/had federations - branches - with some other name)
-   PARTIDO NACIONALISTA VASCO
-   BLOQUE NACIONALISTA GALLEGO
-   CONVERGÈNCIA I UNIÓ
-   UNIDAS PODEMOS - IU (beware that here they have had various names - IU, podem, ezker batua, …- and have not always gone together, but here we will analyze them together)
-   ESQUERRA REPUBLICANA DE CATALUNYA
-   EH - BILDU (are now a coalition of parties formed by Sortu, Eusko Alkartasuna, Aralar, Alternatiba)
-   MÁS PAÍS
-   VOX

The rest of them will be classified as "OTHER".

```{r}
election_data <- election_data |> 
  mutate(
    partido_2 = case_when(
      str_detect(partido, "PARTIDO DE LA CIUDADANIA|PARTIDO DE LA CIUDADANÍA") ~ "CIUDADANOS",
      str_detect(partido, "PSOE|PARTIDO SOCIALISTA|PARTIT DELS SOCIALISTES DE CATALUNYA|PARTIT SOCIALISTA OBRER ESPANYOL|PARTIDO DOS SOCIALISTAS DE GALICIA") ~ "PARTIDO SOCIALISTA OBRERO ESPAÑOL",
      str_detect(partido, "PARTIDO POPULAR") ~ "PARTIDO POPULAR",
      str_detect(partido, "PARTIDO NACIONALISTA VASCO") ~ "PARTIDO NACIONALISTA VASCO",
      str_detect(partido, "BLOQUE NACIONALISTA GALEGO") ~ "BLOQUE NACIONALISTA GALEGO",
      str_detect(partido, "CONVERGENCIA I UNIO|CONVERGÈNCIA I UNIÓ|CONVERGÈNCIA i UNIÓ") ~ "CONVERGÈNCIA I UNIÓ",
      str_detect(partido, "PODEM|IZQUIERDA UNIDA|ESQUERRA UNIDA|ESQUERDA UNIDA|EZKER BATUA|UNIDAD POPULAR") ~ "UNIDAS PODEMOS - IU",
      str_detect(partido, "ESQUERRA REPUBLICANA") ~ "ESQUERRA REPUBLICANA DE CATALUNYA",
      str_detect(partido, "BILDU") ~ "EH-BILDU",
      str_detect(partido, "MÁS PAÍS") ~ "MÁS PAÍS",
      str_detect(partido, "VOX") ~ "VOX",
      TRUE ~ "OTHER"
    )
  )

# IMPORTANT: From now on, remember to use the variable "partido_2" for the analysis.
# I didn't want to delete the original party variable just in case we need it for reference
```

Finally, I will create an additional variable including the acronyms of the parties. I thought about joining the `abbrev` dataset using `left_join()` but there were so many inconsistencies in the different names that it was easier and quicker to just add a new column.

```{r}
election_data <- election_data |> 
  mutate(siglas = case_when(
    partido_2 == "CIUDADANOS" ~ "CS",
      partido_2 == "PARTIDO SOCIALISTA OBRERO ESPAÑOL" ~ "PSOE",
      partido_2 == "PARTIDO POPULAR" ~ "PP",
      partido_2 == "PARTIDO NACIONALISTA VASCO" ~ "EAJ-PNV",
      partido_2 == "BLOQUE NACIONALISTA GALEGO" ~ "BNG",
      partido_2 == "CONVERGÈNCIA I UNIÓ" ~ "CIU",
      partido_2 == "UNIDAS PODEMOS - IU" ~ "UP",
      partido_2 == "ESQUERRA REPUBLICANA DE CATALUNYA" ~ "ERC",
      partido_2 == "EH-BILDU" ~ "EH-BILDU",
      partido_2 == "MÁS PAÍS" ~ "MP",
      partido_2 == "VOX" ~ "VOX",
      TRUE ~ "OTHER"
    )
  ) |> 
  relocate(siglas, .before = partido_2)
```


## surveys

For this dataframe I also need to use `pivot_longer( )` to create a new variable that stores the name of each party, and another one that includes the voting intention for each of them:

```{r}
surveys <- surveys |> 
  pivot_longer(cols = -(type_survey:turnout), 
               names_to = "party_name",
               values_to = "vote_intention")           
```

Now, I remove the `type_survey` variable, because all of them are national surveys, so it does not provide any useful information:

```{r}
surveys <- surveys |> 
  select(-type_survey)
```

Then, I will remove...

-   Surveys before 2008
-   Surveys whose sample size is less than 750 or NA
-   Surveys whose fieldwork days is 1 or less
-   Surveys that are exit_polls (exit_poll = TRUE)

```{r}
# I create a new variable called fieldwork_days:

surveys <- surveys |>
  mutate(fieldwork_days = as.numeric(field_date_to - field_date_from)) |> 
  relocate(fieldwork_days, .after = field_date_to)

# I filter according to the instructions:

surveys <- surveys |> 
  filter(date_elec >= as.Date("2008-01-01"),                  
         size >= 750,            
         !is.na(size),           
         fieldwork_days > 1,            
         exit_poll == FALSE)
```

Finally, I will filter the parties again. But first, I will create an additional variable including the name of the parties (the original dataset only includes acronyms). Again, I also thought about joining the `abbrev` dataset using `left_join()` but there were so many inconsistencies in the different denominations that it was easier and quicker to just add a new column.

```{r}
surveys <- surveys |>
  rename(party_abbrev = party_name) |> 
  mutate(
    party_name = case_when(
      party_abbrev == "CS" ~ "CIUDADANOS",
      party_abbrev == "PSOE" ~ "PARTIDO SOCIALISTA OBRERO ESPAÑOL",
      party_abbrev == "PP" ~ "PARTIDO POPULAR",
      party_abbrev == "EAJ-PNV" ~ "PARTIDO NACIONALISTA VASCO",
      party_abbrev == "BNG" ~ "BLOQUE NACIONALISTA GALLEGO",
      party_abbrev == "CIU" ~ "CONVERGÈNCIA I UNIÓ",
      party_abbrev == "UP" ~ "UNIDAS PODEMOS - IU",
      party_abbrev == "ERC" ~ "ESQUERRA REPUBLICANA DE CATALUNYA",
      party_abbrev == "EH-BILDU" ~ "EH-BILDU",
      party_abbrev == "MP" ~ "MÁS PAÍS",
      party_abbrev == "VOX" ~ "VOX",
      TRUE ~ "OTHER"
    )
  ) |> 
  relocate(party_name, .before = party_abbrev)
```

## Joining dataframes

As a last step, I will create a new variable in the `election_data` dataframe called “cod_mun”. I will do that pasting values from `codigo_ccaa`, `codigo_provincia` and `codigo_municipio` columns and then use it as a key so we are able to join the `cod_mun` dataframe.

```{r}
election_data <- election_data |> 
  mutate(cod_mun = paste(codigo_ccaa, codigo_provincia, codigo_municipio, sep = "-")) |> 
  relocate(cod_mun, .before = codigo_ccaa) |> 
  left_join(cod_mun, by = "cod_mun") |> 
  relocate(municipio, .before = cod_mun)
```

## Export to data/processed

```{r}
saveRDS(election_data, "data/processed/election_data.rds")
saveRDS(surveys, "data/processed/surveys.rds")
```
